interface and rent system almost ready todo post methods, filter

+++ backend/Controllers/ManageController.cs
@@ -133,7 +133,7 @@ namespace Bikepark.Controllers
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!ItemExists(item.ItemID))
+                    if (!ItemExists((int)item.ItemID))
                     {
                         return NotFound();
                     }
